requirements.txt
fastapi==0.104.1
mysql-connector-python==8.2.0
pydantic==2.4.2
pytz==2021.1
python-dotenv==1.0.0
uvicorn==0.24.0
pyyaml==6.0.1

# Testing requirements
pytest==7.0.1
pytest-cov==3.0.0
requests==2.26.0
pytest-mock==3.6.1
pytest-asyncio==0.18.3
pytest-xdist==2.5.0 